services/auth-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic[email]==2.5.0
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
